Changed all data object classes into an interface and a function that returns a plain JS object that matches the interface. This fixes Redux complaining that data objects in the store are not serializable, because classes are not fully serializable. Methods on those classes have be reworked as functions that take an additional object as a parameter.
Also the Item list used in the redux state has been changed to type Item[] instead of (Item|undefined)[] because there is an EmptySlot object to represent when nothing has been selected. This should make typing simpler for this list, and the EmptySlots can be used to modify how many empty slots are displayed in the case that builds go over 6 items.